web/requirements.txt
Django
psycopg2
gunicorn
django-crispy-forms
setuptools
wheels
